requirements.txt
# StartSmart MVP - Python Dependencies
# Phase 0: Contracts, Data Generation, Database Seeding
# Install: pip install -r requirements.txt

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment & Configuration
python-dotenv==1.0.0

# Google Places API (Phase 1)
googlemaps==4.10.0

# Geospatial Operations
shapely==2.0.2

# Development Dependencies
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
